flex controller file

+++ controller/message.go
@@ -26,7 +26,7 @@ func MessageAction(c *gin.Context) {
 
 	if user, exist := usersLoginInfo[token]; exist {
 		userIdB, _ := strconv.Atoi(toUserId)
-		chatKey := genChatKey(user.Id, int64(userIdB))
+		chatKey := genChatKey(int64(user.Id), int64(userIdB))
 
 		atomic.AddInt64(&messageIdSequence, 1)
 		curMessage := Message{
@@ -53,7 +53,7 @@ func MessageChat(c *gin.Context) {
 
 	if user, exist := usersLoginInfo[token]; exist {
 		userIdB, _ := strconv.Atoi(toUserId)
-		chatKey := genChatKey(user.Id, int64(userIdB))
+		chatKey := genChatKey(int64(user.Id), int64(userIdB))
 
 		c.JSON(http.StatusOK, ChatResponse{Response: Response{StatusCode: 0}, MessageList: tempChat[chatKey]})
 	} else {
